Split out in separate component
Props @marekhrabe.
