+++ packages/components/src/date-time/time.js
@@ -10,14 +10,13 @@ import moment from 'moment';
  */
 import { Component } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
-import { __experimentalGetSettings as getDateSettings } from '@wordpress/date';
 
 /**
  * Internal dependencies
  */
 import Button from '../button';
 import ButtonGroup from '../button-group';
-import Tooltip from '../tooltip';
+import TimeZone from './timezone';
 
 /**
  * Module Constants
@@ -279,34 +278,6 @@ class TimePicker extends Component {
 	render() {
 		const { is12Hour } = this.props;
 		const { year, minutes, hours, am } = this.state;
-
-		const TimeZone = () => {
-			const { timezone } = getDateSettings();
-			const offset =
-				timezone.offset > 0 ? '+' + timezone.offset : timezone.offset;
-			const zoneAbbr = moment
-				.tz( new Date(), timezone.string )
-				.zoneAbbr();
-			const timezoneDetail =
-				'UTC' === timezone.string
-					? __( 'Coordinated Universal Time' )
-					: '(UTC' +
-					  offset +
-					  ') ' +
-					  timezone.string.replace( '_', ' ' );
-			const timezoneAbbr =
-				'' !== timezone.string && isNaN( zoneAbbr )
-					? zoneAbbr
-					: 'UTC' + offset;
-
-			return (
-				<Tooltip position="top center" text={ timezoneDetail }>
-					<div className="components-datetime__timezone">
-						{ timezoneAbbr }
-					</div>
-				</Tooltip>
-			);
-		};
 
 		return (
 			<div className={ classnames( 'components-datetime__time' ) }>
